requirements.txt
requests==2.28.1
paramiko==3.1.0
python-dotenv==1.0.0
